Updated with changes to Convertor page. Also moved a lot of that code to a Container component since I realized I had the component on the page file
